Remove activity counter internal state from public interface
This PR removes the activity counter internal state from the public
interface and moves the code into the internet identity crate.

This PR only _moves_ code. It will be followed by additional refactorings
to separate the counting infrastructure more from the things being counted.
