+++ src/internet_identity/internet_identity.did
@@ -170,8 +170,6 @@ type InternetIdentityStats = record {
     };
     archive_info: ArchiveInfo;
     canister_creation_cycles_cost: nat64;
-    active_anchor_stats: opt ActiveAnchorStatistics;
-    domain_active_anchor_stats: opt DomainActiveAnchorStatistics;
     max_num_latest_delegation_origins: nat64;
     latest_delegation_origins: vec FrontendHostname
 };
@@ -207,60 +205,6 @@ type RateLimitConfig = record {
     time_per_token_ns : nat64;
     // How many tokens are at most generated (to accommodate peaks).
     max_tokens: nat64;
-};
-
-type ActiveAnchorStatistics = record {
-    // Stats for the last completed collection period for daily and monthly active anchors
-    completed: CompletedActiveAnchorStats;
-    // ongoing periods for daily and monthly active anchors
-    ongoing: OngoingActiveAnchorStats;
-};
-
-type CompletedActiveAnchorStats = record {
-    daily_active_anchors: opt ActiveAnchorCounter;
-    monthly_active_anchors: opt ActiveAnchorCounter;
-};
-
-type OngoingActiveAnchorStats = record {
-    // Ongoing active anchor counter for the current 24 h time bucket.
-    daily_active_anchors: ActiveAnchorCounter;
-    // Monthly active users are collected using 30-day sliding windows.
-    // This vec contains up to 30 30-day active windows each offset by one day.
-    // The vec is sorted, new collection windows are added at the end.
-    monthly_active_anchors: vec ActiveAnchorCounter;
-};
-
-type ActiveAnchorCounter = record {
-    start_timestamp: Timestamp;
-    counter: nat64;
-};
-
-type DomainActiveAnchorStatistics = record {
-    // Stats for the last completed collection period for daily and monthly active anchors
-    completed: DomainCompletedActiveAnchorStats;
-    // ongoing periods for daily and monthly active anchors
-    ongoing: DomainOngoingActiveAnchorStats;
-};
-
-type DomainCompletedActiveAnchorStats = record {
-    daily_active_anchors: opt DomainActiveAnchorCounter;
-    monthly_active_anchors: opt DomainActiveAnchorCounter;
-};
-
-type DomainOngoingActiveAnchorStats = record {
-    // Ongoing active anchor counter for the current 24 h time bucket.
-    daily_active_anchors: DomainActiveAnchorCounter;
-    // Monthly active users are collected using 30-day sliding windows.
-    // This vec contains up to 30 30-day active windows each offset by one day.
-    // The vec is sorted, new collection windows are added at the end.
-    monthly_active_anchors: vec DomainActiveAnchorCounter;
-};
-
-type DomainActiveAnchorCounter = record {
-    start_timestamp: Timestamp;
-    ic0_app_counter: nat64;
-    internetcomputer_org_counter: nat64;
-    both_ii_domains_counter: nat64;
 };
 
 // Init arguments of II which can be supplied on install and upgrade.
