+++ src/internet_identity/src/anchor_management.rs
@@ -1,5 +1,5 @@
-use crate::active_anchor_stats::IIDomain;
 use crate::archive::{archive_operation, device_diff};
+use crate::ii_domain::IIDomain;
 use crate::state::RegistrationState::DeviceTentativelyAdded;
 use crate::state::TentativeDeviceRegistration;
 use crate::storage::anchor::{Anchor, Device};
