+++ src/internet_identity_interface/src/internet_identity.rs
@@ -7,9 +7,6 @@ pub mod types;
 /// Helpful data conversions for the types.
 pub mod conversions;
 
-/// Anchor activity counter trait and implementations thereof.
-pub mod anchor_activity_counter;
-
 /// Helpful implementations on the authentication method data type.
 pub mod authn_method;
 /// Helpful implementations on the device data type.
